backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.0.1
aiosmtplib==3.0.1
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8